Fixed byte_counter range issues and TAG issues
Fixed an issue where the byte 15 wasn't being read. increased the cap to 16, and everything works now.

Also set the tag in cache at the end of the allocate portion, that way, when doing write_back, the cache block should contain the tag we need to get the address of the write_back in memory.